added chakra
