+++ src/App.tsx
@@ -1,10 +1,11 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import "./App.css";
-import Login from "./components/Login.tsx";
+// import Login from "./components/Login.tsx";
 import MainPage from "./components/MainPage.tsx";
 import Records from "./components/Records.tsx";
 import ErrorPage from "./components/ErrorPage.tsx";
+import Signup from "./components/Signup.tsx";
 
 function App() {
   return (
@@ -12,7 +13,7 @@ function App() {
       <Router>
         <Routes>
           <Route path="/" element={<MainPage />} />
-          <Route path="/login" element={<Login />} />
+          <Route path="/login" element={<Signup />} />
           <Route path="/records" element={<Records />} />
           <Route path="*" element={<ErrorPage />} />
         </Routes>
